🔨 REFACTOR: migrate PostTrackback to functional component with TS support
